Replace cnijfilter-mp620 by cnijfilter-mp630
I own an MP630 and in v3.00 there is custom code for this printer model. So use
it instead of the MP610/620 stuff. Also use the original PPD file. This may
change again if I see what are the benefits of the other PPD.

This makes the package much simpler because no v2.80 is needed anymore and we
can get rid of a lot of patches.