kodi-platform: lock version to 45d6ad1
Newer versions require a current libplatform. This in itself it not a problem
but it breaks all current kodi pvr addons. Patches for these addons are still
in work or under review so for now we just stick to an old version.

This will be fixed as soon as kodi plugins start building against a current
version of libplatform.