openafs-modules-dkms: do not attempt to load MAKEFLAGS from makepkg.conf
The current version of dkms.conf causes issues with pacman's dkms hook
if MAKEFLAGS are not set in makepkg.conf. While it would be possible
to work around this issue, hijacking makepkg.conf like this is probably
not a good idea anyway. If people want to use certain flags for all
build processes, they can globally export MAKEFLAGS in their
environment.

Thanks to seally1186 for tracking down this issue!