jabref: force use of Java version 21
Regardless of $JAVA_HOME, gradle will try to find a JDK with the version
specified as toolchain->languageVersion. This lets the build fail if no
JDK 20 is installed.

For now, stick to a single valid JDK version. The best option is JDK 21
which is currently available as jdk-openjdk in the Arch repositories.

Thanks a lot to Pavlov and vanja_z, providing all the information
required to fix this issue.