jabref: make sure to launch JabRef with installed JDK
Arch Linux's official JDK and JRE packages of the same version conflict
with each other. Hence, depending on java-environment=21 works in the
majority of cases. However, there are JDK packages in the AUR that
provide java-environment but not jdk-openjdk and hence do not conflict
with JRE packages. Make sure to launch JabRef with a full JDK and don't
rely on it being chosen automatically.

Fixes a7954604dca52792cf7f6cac0bfe832d806a4a35